Populate pages and images
Pages were created and images were added to the recipes pages. No
issues here. Though the images file for sushi is .avif so I am not
sure how that will turn out.

This is also an attempt by me to use best practices for git commit
messages. Some key things to consider are: 50 characters for the
subject. Wrapping the body at ~75 characters. Using the body to explain
what and why vs. how.